Fix u reward wrapper

+++ rl/notebooks/understand_hyperparameters.py
@@ -123,8 +123,9 @@ def save_best_policy(trainer, episodes):
         os.makedirs(os.path.join(trainer.global_vars['reporter_dir'], 'best'), exist_ok=True)
         trainer.global_vars['best_eval_reward'] = mean_r
         # save policy
-        shutil.rmtree(os.path.join(trainer.global_vars['reporter_dir'], 'best', 'policy'), ignore_errors=True)
-        trainer.get_policy().export_model(os.path.join(trainer.global_vars['reporter_dir'], 'best', 'policy'))
+        if not trainer.global_vars['unbalance']:
+            shutil.rmtree(os.path.join(trainer.global_vars['reporter_dir'], 'best', 'policy'), ignore_errors=True)
+            trainer.get_policy().export_model(os.path.join(trainer.global_vars['reporter_dir'], 'best', 'policy'))
         # save plots
         ep = episodes[-1]
         f = plot_new(ep.hist_data['tracking'], trainer.iteration, trainer.global_vars['unbalance'])
